Create WizardTemplateDetails and DefaultTemplateDetails
